Finish enums for word bounds.
Previously, all word boundary characters were dropped and
replaced with a single whitespace. The structure of the
words was changed from a simple String, to an enum with
two types: Word and Separator to allow differential treatment.
The program now preserves all word-boundary characters.

Additionally, fixed a bug in which vowel detection was unable
to detect capital letters.